Implement cluster label support
This commit does two major things:

(1) In case cluster labels are long, we first expand the margins
of all enclosed nodes such that we can add our labels later with
adequate space. Unfortunately this needs to be done early enough
for us to position the nodes accordingly.

(2) We must make enough space for the label at each level of the
cluster. This requires recursively computing the necessary padding.